@@ -7,35 +7,32 @@
     "homepage": "./",
     "main": "public/electron.js",
     "dependencies": {
-        "@electron/remote": "^1.2.0",
         "@fortawesome/fontawesome-svg-core": "^1.2.35",
         "@fortawesome/free-solid-svg-icons": "^5.15.3",
-        "@fortawesome/react-fontawesome": "^0.1.14",
-        "@tippyjs/react": "^4.2.5",
-        "ajv": "^8.6.2",
-        "electron-store": "^8.0.0",
-        "electron-updater": "^4.6.0",
-        "electron-util": "^0.17.2",
+        "@fortawesome/react-fontawesome": "^0.1.16",
+        "@tippyjs/react": "^4.2.6",
+        "ajv": "^8.8.0",
+        "electron-store": "^8.0.1",
+        "electron-updater": "^4.6.2",
         "electron-window-state": "^5.0.3",
-        "filesize": "^6.3.0",
+        "filesize": "^8.0.6",
         "fs-extra": "^10.0.0",
         "i18n": "file:./i18n",
         "md5": "^2.3.0",
         "react": "^17.0.1",
         "react-dom": "^17.0.2",
-        "react-router-dom": "^5.2.0",
+        "react-router-dom": "^6.0.2",
         "react-scripts": "^4.0.3",
-        "tar": "^6.1.10",
+        "tar": "^6.1.11",
         "tus-js-client": "^2.3.0"
     },
     "devDependencies": {
-        "concurrently": "^6.2.1",
+        "concurrently": "^6.4.0",
         "cross-env": "^7.0.3",
-        "electron": "^13.5.1",
-        "electron-builder": "^22.11.7",
-        "electron-notarize": "^1.1.0",
+        "electron": "^15.3.1",
+        "electron-builder": "^22.13.1",
+        "electron-notarize": "^1.1.1",
         "node-sass": "^6.0.1",
-        "typescript": "^4.4.2",
         "wait-on": "^6.0.0"
     },
     "scripts": {
@@ -85,6 +82,6 @@
         ]
     },
     "browserslist": [
-        "electron 11.0"
+        "electron 15.0"
     ]
 }
